+++ ruoyi-admin/src/main/java/com/ruoyi/data/service/SubscribeService.java
@@ -18,7 +18,7 @@ public interface SubscribeService {
      * @param id 订阅关系ID
      * @return 订阅关系
      */
-    public Subscribe selectSubscribeById(Integer id);
+    public Subscribe selectSubscribeById(Long id);
 
     /**
      * 查询订阅关系列表
@@ -58,7 +58,7 @@ public interface SubscribeService {
      * @param id 订阅关系ID
      * @return 结果
      */
-    public int deleteSubscribeById(Integer id);
+    public int deleteSubscribeById(Long id);
 
     int insertBatch(List<Subscribe> list);
